Model data off of states not query parameters.
Previously states based on query parameters worked only if a single
parameter was updated at a time. While that's fine for user
interactions, it's not great for setting defaults when the JSON file
changes. For example, we'd like to be able to set a defult value for
start and end timestamps based on the events in the JSON file. But that
creates a race where we write some query parameters and not others.

Now we just write every value to a state and use useEffect to update all
query parameters at once so things remain consistent. We use an
additional event to track the compressed parsed data because compression
is asynchronous.